also convert sample_weights to float32 when calling sklearn

+++ src/metrax/nnx/nnx_wrapper_test.py
@@ -91,7 +91,7 @@ class NnxWrapperTest(parameterized.TestCase):
     expected = sklearn_metrics.mean_squared_error(
         y_true.astype('float32').flatten(),
         y_pred.astype('float32').flatten(),
-        sample_weight=sample_weights.flatten(),
+        sample_weight=sample_weights.astype('float32').flatten(),
     )
     # Use lower tolerance for lower precision dtypes.
     rtol = 1e-2 if y_true.dtype in (jnp.float16, jnp.bfloat16) else 1e-05
